trying to compile on win64

+++ src/object_loader/read_color.c
@@ -7,8 +7,8 @@ int			read_color(t_obj_data *data, const t_token *token)
   if (!check_tokens_number(token, 2))
 		read_object_error("A texture needs two arguments.");
 
-  new_color.x = atof(&token->cursor[0]);
-	new_color.y = atof(&token->cursor[1]);
+  new_color.x = (float)atof(&token->cursor[0]);
+	new_color.y = (float)atof(&token->cursor[1]);
 
 	ft_lstadd(&data->uvs, ft_lstnew(&new_color, sizeof(t_vec2)));
